fix rewrite integrated test
