+++ shenyu-integrated-test/pom.xml
@@ -39,6 +39,7 @@
         <module>shenyu-integrated-test-motan</module>
         <module>shenyu-integrated-test-sofa</module>
         <module>shenyu-integrated-test-websocket</module>
+        <module>shenyu-integrated-test-rewrite</module>
         <module>shenyu-integrated-test-combination</module>
         <module>shenyu-integrated-test-sdk-apache-dubbo</module>
         <module>shenyu-integrated-test-sdk-alibaba-dubbo</module>
